pyproject.toml
[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[project]
name = "context"
version = "0.1.0"
description = "Node-based image processing desktop application built with PyQt6."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.0.0",
    "PyQt6>=6.9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
context = ["ui/translations.json"]

[project.urls]
Homepage = "https://github.com/RafaelCasamaximo/contExt/"
Repository = "https://github.com/RafaelCasamaximo/contExt.git"
Documentation = "https://github.com/RafaelCasamaximo/contExt/blob/main/README.md"
Issues = "https://github.com/RafaelCasamaximo/contExt/issues/"
Changelog = "https://github.com/RafaelCasamaximo/contExt/commits/main/"
